pyproject.toml
[project]
name = "docauth"
version = "2.0.0"
description = "Document Forgery Detection and Analysis"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
keywords = ["forensics", "forgery-detection", "deep-learning", "computer-vision", "ocr"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]
dependencies = [
    # Core ML
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "tensorflow>=2.15.0",
    "timm>=1.0.0",
    "transformers>=4.45.0",
    "pytorch-metric-learning>=2.3.0",
    "accelerate>=0.34.0",
    # Image processing
    "Pillow>=10.4.0",
    "opencv-python>=4.10.0",
    "numpy>=1.26.4",
    "scipy>=1.13.1",
    "scikit-learn>=1.5.2",
    "scikit-image>=0.24.0",
    "PyWavelets>=1.7.0",
    "imageio>=2.34.0",
    "albumentations>=1.4.0",
    # OCR
    "easyocr>=1.7.1",
    # Streamlit UI
    "streamlit>=1.40.0",
    "plotly>=5.22.0",
    "streamlit-image-comparison>=0.0.4",
    # API Backend
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.12",
    "pydantic>=2.9.0",
    # Utilities
    "tqdm>=4.66.5",
    "python-dotenv>=1.0.0",
    "huggingface-hub>=0.24.0",
    "requests>=2.31.0",
    # Notebooks
    "jupyter>=1.1.0",
    "ipywidgets>=8.1.0",
    "matplotlib>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]
gpu = [
    # Install PyTorch with CUDA support separately:
    # uv add torch --index https://download.pytorch.org/whl/cu124
]

[project.urls]
Homepage = "https://github.com/trinity652/DocAuth"
Repository = "https://github.com/trinity652/DocAuth"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
